+++ src/MNoteProvider.ClientService/SignalRClient/NoteEventRelay.cs
@@ -1,12 +1,10 @@
 using MNoteProvider.ClientService.Abstractions;
-using MNoteProvider.Common.Abstractions.DTOs;
 using MNoteProvider.Common.DTOs;
 namespace MNoteProvider.ClientService.SignalRClient;
 
 /// <summary>Relays note hub events as client-side note notifications.</summary>
 internal sealed class NoteEventRelay : INoteEventRelay
 {
-
 
     /// <inheritdoc/>
     public event Action<NoteDto>? NoteCreatedNotification;